requirements.txt
alembic==1.0.6
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
oauthlib==2.1.0
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
requests==2.21.0
requests-oauthlib==1.1.0
six==1.12.0
SQLAlchemy==1.2.16
urllib3==1.24.1
Werkzeug==0.14.1
